# perf: use all processors for faster rendering