This is more of a suggestion. I had to break many things to make it compile.

Some timing results:

Before:

```
cargo run --release -- -f scene.yml  87,48s user 0,04s system 99% cpu 1:27,52 total
cargo run --release -- -f scene.yml  87,22s user 0,12s system 99% cpu 1:27,55 total
cargo run --release -- -f scene.yml  88,17s user 0,08s system 99% cpu 1:28,26 total
```

After:

```
cargo run --release -- -f scene.yml  135,41s user 0,81s system 690% cpu 19,724 total
cargo run --release -- -f scene.yml  136,18s user 0,84s system 704% cpu 19,452 total
cargo run --release -- -f scene.yml  135,47s user 0,77s system 716% cpu 19,013 total
```